Sketch of use_tags functionality.
A role inclusion may specify use_tags: [a, b, c] to *set* the tags
that will be used from the included role.

This brings tag use closer to "method selection" and provides a
playbook facility that everyone I've canvassed expects to be there
in the first place: the programmatic selection of tags to run.